Code refactored + comments
